+++ routes/__init__.py
@@ -1,0 +1,7 @@
+from fastapi import APIRouter
+
+router = APIRouter()
+
+@router.get("/ping")
+def ping():
+    return {"status": "ok"}
